SPARE and SPACE circuit PanelSchedule change
The issue is that if one tries to snoop current selection while selecting a "spare" or "space" circuit in a panel schedule the application throws an error because get_elements() cannot be called on category BuiltInCategory.OST_ElectricalInternalCircuits. 

The workaround here is to check if the object category is OST_ElectricalInternalCircuits; if it is then don't try to add Elements because you can't, if it isn't then add elements.