docs: update Readme  - add instructions on how to run
Thank you for maintaining this library!

The following PR makes it explicitly clear for those new to the extension - how to run it. It may seem obvious to Revit experts. But it was not obvious to me.

Feel free to discard this PR if it doesn't add value. It is merely a suggestion.